--- test-tests/shared/src/main/scala/zio/test/AsyncBaseSpec.scala
@@ -1,7 +1,0 @@
-package zio.test
-
-import zio.DefaultRuntime
-
-trait AsyncBaseSpec extends DefaultRuntime {
-  def run: List[Async[(Boolean, String)]]
-}
